Added license information and fleshed out the about page
Moved all the info in the about page to the page instead of a data file,
it just feels silly to write it in there instead of on the page itself.

Posts are licensed under CC-BY-SA 4.0 and all code under MIT. There are
links in the footer to the respective licenses. CC-BY-SA 4.0 links to
the CC site whereas the MIT license shows a modal containing the text.